+++ tests/Handler/CurlFactoryTest.php
@@ -117,6 +117,64 @@ class CurlFactoryTest extends TestCase
         self::assertEquals(10, $_SERVER['_curl'][\CURLOPT_LOW_SPEED_LIMIT]);
     }
 
+    public function testRejectsCurlProxyHeaderEntriesContainingNewlines(): void
+    {
+        $proxyHeaderOption = self::proxyHeaderOption();
+        $factory = new CurlFactory(3);
+
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('CURLOPT_PROXYHEADER');
+
+        $factory->create(new Psr7\Request('GET', 'http://example.com'), [
+            'curl' => [
+                $proxyHeaderOption => ["X-Decoy: v\r\nProxy-Authorization: Basic abc"],
+            ],
+        ]);
+    }
+
+    public function testRejectsCurlHttpHeaderEntriesContainingNewlines(): void
+    {
+        $conf = [\CURLOPT_HTTPHEADER => ["X-Decoy: v\r\nProxy-Authorization: Basic abc"]];
+
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('CURLOPT_HTTPHEADER');
+
+        self::normalizeCurlHeaderOptions($conf);
+    }
+
+    public function testNormalizesStringableCurlHeaderEntriesBeforeProxyTunnelSignature(): void
+    {
+        $proxyHeaderOption = self::proxyHeaderOption();
+        $conf = [
+            \CURLOPT_PROXY => 'http://proxy.example.com:8080',
+            $proxyHeaderOption => [new class() {
+                public function __toString(): string
+                {
+                    return 'Proxy-Authorization: Basic abc';
+                }
+            }],
+        ];
+
+        self::normalizeCurlHeaderOptions($conf);
+
+        self::assertSame(['Proxy-Authorization: Basic abc'], $conf[$proxyHeaderOption]);
+        $delegated = self::computeProxyTunnelSignature('8.20.0', 'https://example.com', [
+            \CURLOPT_PROXY => 'http://proxy.example.com:8080',
+        ]);
+        $literalHeader = self::computeProxyTunnelSignature('8.20.0', 'https://example.com', $conf);
+        self::assertNotNull($literalHeader);
+        self::assertNotSame($delegated, $literalHeader);
+    }
+
+    public function testLeavesNormalCurlHeaderEntriesUnchanged(): void
+    {
+        $conf = [\CURLOPT_HTTPHEADER => ['Accept: application/json']];
+
+        self::normalizeCurlHeaderOptions($conf);
+
+        self::assertSame(['Accept: application/json'], $conf[\CURLOPT_HTTPHEADER]);
+    }
+
     public function testAppliesConfiguredCurlShareHandle(): void
     {
         self::skipIfCurlShareIsUnavailable();
@@ -2570,6 +2628,19 @@ class CurlFactoryTest extends TestCase
         }
 
         return (int) \constant('CURLOPT_PROXYHEADER');
+    }
+
+    /**
+     * @param array<int|string, mixed> $conf
+     */
+    private static function normalizeCurlHeaderOptions(array &$conf): void
+    {
+        $method = new \ReflectionMethod(CurlFactory::class, 'normalizeCurlHeaderOptions');
+        if (\PHP_VERSION_ID < 80100) {
+            $method->setAccessible(true);
+        }
+
+        $method->invokeArgs(null, [&$conf]);
     }
 
     private static function redactProxyUserInfo(string $error, ?string $proxy): string